+++ app/src/main/java/fi/thl/koronahaavi/di/AppModule.kt
@@ -2,22 +2,15 @@ package fi.thl.koronahaavi.di
 
 import android.content.Context
 import android.content.SharedPreferences
-import androidx.security.crypto.EncryptedSharedPreferences
-import androidx.security.crypto.EncryptedSharedPreferences.PrefKeyEncryptionScheme
-import androidx.security.crypto.EncryptedSharedPreferences.PrefValueEncryptionScheme
-import androidx.security.crypto.MasterKeys
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ApplicationComponent
 import dagger.hilt.android.qualifiers.ApplicationContext
 import fi.thl.koronahaavi.BuildConfig
-import fi.thl.koronahaavi.R
-import fi.thl.koronahaavi.data.DefaultExposureRepository
-import fi.thl.koronahaavi.data.ExposureDao
-import fi.thl.koronahaavi.data.ExposureRepository
-import fi.thl.koronahaavi.data.KeyGroupTokenDao
+import fi.thl.koronahaavi.data.*
 import fi.thl.koronahaavi.service.UserAgentInterceptor
+import fi.thl.koronahaavi.settings.UserPreferences
 import okhttp3.CertificatePinner
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
@@ -34,18 +27,16 @@ annotation class DatabaseName
 @InstallIn(ApplicationComponent::class)
 object AppModule {
 
-    const val SHARED_PREFERENCES_NAME = "fi.thl.koronavilkku.prefs"
+    @Singleton
+    @Provides
+    fun provideUserPreferences(@ApplicationContext context: Context): UserPreferences {
+        return UserPreferences(context)
+    }
 
     @Singleton
     @Provides
     fun provideSharedPreferences(@ApplicationContext context: Context): SharedPreferences {
-        return EncryptedSharedPreferences.create(
-            SHARED_PREFERENCES_NAME,
-            MasterKeys.getOrCreate(MasterKeys.AES256_GCM_SPEC), // gets encryption key alias from keystore
-            context,
-            PrefKeyEncryptionScheme.AES256_SIV,
-            PrefValueEncryptionScheme.AES256_GCM
-        )
+        return EncryptedSharedPreferencesWrapper(context)
     }
 
     // this allows database testing with a different name so it does not interfere with actual app db